fixed tile collision, sprite manager set up
tile collision now works with the scrolling map, scrolling is now handled by the platform player and the sprite manager. the player notes if they are moving past a scrolling threshhold and sends they're velocity to the sprite manager which moves every registered sprite, including the player. it also records the offset caused by scrolling which is passed into the collision manager to account for the tiles being in a different place compared to their spot in the m_level array.
also moved the level map reading and loading back into the engine and made a function, loadlevel(), which tells the engine which level to load.
also made it so kiki flips their sprite when moving backwards